refactor(#273): GameScorerServiceImpl SRP 분리 — 5개 책임별 서비스로 분할
821줄 모놀리식 GameScorerServiceImpl을 단일 책임 원칙에 따라 5개 서비스로 분리:
- GameLifecycleService: 경기 시작/종료/몰수/취소/이닝 전환
- PlateAppearanceRecordService: 타석 결과 기록
- GameUndoService: 이벤트 되돌리기
- BaseRunningRecordService: 주루 기록
- GameSubstitutionService: 선수 교체
